V 0.6.0
-Line2 Class:
Solved a bug in the line collisionWith circle method and line collisionPoint method.
Optimized method line collisionWith circle and method line collisionPoint circle.

-Input Class:
Now there's a method to get the code of the last (keyboard, mouse or mouse wheel) pressed or released.
Method to replace codes of an Action.
Method to remove codes of an Action.
Method to add codes to an Action.
Method to remove all codes of an Action.

-Vector2
Implemented Clone(), HashCode() and Equals().